Split SvgColourMatrix
